Gathered/collected code and cleaned up some imports
