--- frontend/pages/api/datasets.js
@@ -1,7 +1,0 @@
-
-
-export default (req, res) => {
-    res.statusCode = 200
-    res.json(
-        { datasets: { title: "req.title", description: 'test'}})
-  }
