+++ frontend/pages/index.js
@@ -10,12 +10,12 @@ import FilterCategory from '../Components/Filters/FilterCategory'
 import FilterTag from '../Components/Filters/FilterTag'
 
 import { useRouter } from 'next/router'
-import { parseCookies } from '../utils/parseCookies'
+import {PageRender} from './api/serverSideProps';
 
 import {useState,useEffect} from 'react'
 
 // Home page, I think this can be the Data catalogue, just change the name from home to datacatalogue or something
-export default function Home({ data, prevLoggedIn = false, prevLoggedUsername = "", prevPublisherId = "-1", prevUserId = "-1" }) {
+export default function Home() {
   const router = useRouter();
 
 
@@ -90,42 +90,11 @@ export default function Home({ data, prevLoggedIn = false, prevLoggedUsername = 
   )
 }
 
-// ALERT: This ships HTTPS validation and should not be used when we are handling personal information and authentication etc.
-function createRequestOptions(skipHttpsValidation) {
-  const isNode = typeof window === 'undefined';
-  if (isNode) {
-    var Agent = (require('https')).Agent;
-    return {
-      agent: new Agent({ rejectUnauthorized: !skipHttpsValidation })
-    };
-  }
-}
+
 
 // This gets called on every request
-export async function getServerSideProps({ req }) {
-  // Fetch data from external API
-  // Should be changed to host link when this is done, not localhost.
-  const uri = 'https://localhost:5001/api/datasets';
-  const res = await fetch(uri, createRequestOptions(true))
-  const data = await res.json()
-
-  const cookies = parseCookies(req)
-
-  let propsData = { props: { data } }
-
-  if (JSON.stringify(cookies) !== "{}") {
-    propsData = {
-      props: {
-        data,
-        prevLoggedIn: cookies.prevLoggedIn,
-        prevLoggedUsername: cookies.prevLoggedUsername,
-        prevPublisherId: cookies.prevPublisherId,
-        prevUserId: cookies.prevUserId
-      }
-    }
-  }
-  console.log(cookies)
-
+export async function getServerSideProps(context) {
+  const propsData = PageRender("index", context)
   return propsData
