--- frontend/utils/parseCookies.js
@@ -1,5 +1,0 @@
-import cookie from "cookie";
-
-export function parseCookies(req) {
-  return cookie.parse(req ? req.headers.cookie || "" : document.cookie);
-}
